Restructure the qgis application
