--- qgis/commands/__init__.py
@@ -1,7 +1,0 @@
-from .base import QGISCommand
-from .raster_calculation import RasterCalculatorCommand
-
-__all__ = [
-    "QGISCommand",
-    "RasterCalculatorCommand",
-]
