--- qgis/objects/__init__.py
@@ -1,5 +1,0 @@
-from .tree import TreePoint
-
-__all__ = [
-    "TreePoint",
-]
