+++ qgis/src/api/models/__init__.py
@@ -1,0 +1,5 @@
+from .point import Point
+
+__all__ = [
+    "Point",
+]
