+++ qgis/src/api/models/point.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel
+
+class Point(BaseModel):
+    x: float
+    y: float
