+++ qgis/src/api/requests/__init__.py
@@ -1,0 +1,5 @@
+from .burn_request import BurnRequest
+
+__all__ = [
+    "BurnRequest",
+]
