+++ qgis/src/api/requests/burn_request.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel
+from typing import List
+from src.api.models.point import Point
+
+class BurnRequest(BaseModel):
+    points: List[Point]
