+++ qgis/src/configs/__init__.py
@@ -1,0 +1,5 @@
+from .preflight import init_qgis
+
+__all__ = [
+    "init_qgis",
+]
