+++ qgis/src/services/__init__.py
@@ -1,0 +1,5 @@
+from .raster_service import burn_points_to_raster
+
+__all__ = [
+    "burn_points_to_raster",
+]
